+++ Kernel/System/DB.pm
@@ -177,7 +177,7 @@ sub new {
     # check/get extra database configuration options
     # (overwrite with params)
     for (
-        qw(Type Limit DirectBlob Attribute QuoteSingle QuoteBack Connect Encode CaseInsensitive LcaseLikeInLargeText)
+        qw(Type Limit DirectBlob Attribute QuoteSingle QuoteBack Connect Encode CaseSensitive LcaseLikeInLargeText)
         )
     {
         if ( defined $Param{$_} ) {
@@ -1223,11 +1223,11 @@ sub QueryCondition {
 
 # check if database supports LIKE in large text types
 # the first condition is a little bit opaque
-# CaseInsensitive of the database defines, if the database handles case sensitivity or not
+# CaseSensitive of the database defines, if the database handles case sensitivity or not
 # and the parameter $CaseSensitive defines, if the customer database should do case sensitive statements or not.
 # so if the database dont support case sensitivity or the configuration of the customer database want to do this
 # then we prevent the LOWER() statements.
-                    if ( $Self->GetDatabaseFunction('CaseInsensitive') || $CaseSensitive ) {
+                    if ( !$Self->GetDatabaseFunction('CaseSensitive') || $CaseSensitive ) {
                         $SQLA .= "$Key $Type '$Word'";
                     }
                     elsif ( $Self->GetDatabaseFunction('LcaseLikeInLargeText') ) {
@@ -1260,11 +1260,11 @@ sub QueryCondition {
 
 # check if database supports LIKE in large text types
 # the first condition is a little bit opaque
-# CaseInsensitive of the database defines, if the database handles case sensitivity or not
+# CaseSensitive of the database defines, if the database handles case sensitivity or not
 # and the parameter $CaseSensitive defines, if the customer database should do case sensitive statements or not.
 # so if the database dont support case sensitivity or the configuration of the customer database want to do this
 # then we prevent the LOWER() statements.
-                    if ( $Self->GetDatabaseFunction('CaseInsensitive') || $CaseSensitive ) {
+                    if ( !$Self->GetDatabaseFunction('CaseSensitive') || $CaseSensitive ) {
                         $SQLA .= "$Key $Type '$Word'";
                     }
                     elsif ( $Self->GetDatabaseFunction('LcaseLikeInLargeText') ) {