+++ Kernel/System/DB/postgresql_before_8_2.pm
@@ -33,7 +33,7 @@ sub LoadPreferences {
     $Self->{'DB::QuoteSemicolon'}       = '\\';
     $Self->{'DB::QuoteUnderscoreStart'} = '\\\\';
     $Self->{'DB::QuoteUnderscoreEnd'}   = '';
-    $Self->{'DB::CaseInsensitive'}      = 0;
+    $Self->{'DB::CaseSensitive'}        = 1;
     $Self->{'DB::LikeEscapeString'}     = '';
 
     # dbi attributes